+++ src/utils/index.js
@@ -4,7 +4,6 @@
  * @public
  */
 module.exports.exitHandler = require("./exit-handler");
-module.exports.notifHandler = require("./notif-handler");
 module.exports.smartNotifierHandler = require("./smart-notifier-handler");
 module.exports.oAuthHandler = require("./oauth-handler");
 module.exports.actionHandler = require("./action-handler");
@@ -15,7 +14,6 @@ module.exports.staticRouter = require("./static-router");
 module.exports.tokenMiddleware = require("./token-middleware");
 module.exports.requireHullMiddleware = require("./require-hull-middleware");
 module.exports.responseMiddleware = require("./response-middleware");
-module.exports.notifMiddleware = require("./notif-middleware");
 module.exports.smartNotifierMiddleware = require("./smart-notifier-middleware");
 module.exports.smartNotifierErrorMiddleware = require("./smart-notifier-error-middleware");
 module.exports.segmentsMiddleware = require("./segments-middleware");