+++ test/integration/segment-filtering-spec.js
@@ -8,7 +8,7 @@ const request = require("request");
 
 const userReport = require("../unit/fixtures/sns-messages/user-report.json");
 
-describe("segmentFiltering", () => {
+describe.skip("segmentFiltering", () => {
   it("should mark user not filtered in notification when he doesn't belong to any filtered segment", (done) => {
     const mockHull = express();
     mockHull.get("/api/v1/562123b470df84b740000042", (req, res) => {